Fix speech api (#91)
* Changed configuration and socketio directories

* Moved tmpback to new speechback folder, just for speech recognition

* Made socketio use url in config instead of hardcoded URL
