--- frontend/js/connect-socketio.js
@@ -1,1 +1,0 @@
-window.socket = io.connect('http://localhost:8080');
