+++ plugins/devfest/devfest.html
@@ -1,4 +1,6 @@
 <link rel="import" href="/plugins/shared-styles/shared-styles.html">
+<script src="./socketio.js"></script>
+
 <dom-module id="devfest-item">
   <template>
     <style include="shared-styles"></style>
@@ -57,6 +59,8 @@
             speaker: Object,
             speakers: Array,
             speakers_url: String,
+            socket: String,
+            socket_url: String,
             talks: Array,
             talks_url: String
           }
@@ -65,8 +69,10 @@
         connectedCallback () {
           super.connectedCallback();
 
-          socket.on('devfest', (req) => this._handleRequest(req));
-          socket.on('close', (req) => this._showEverything());
+          this.socket = io.connect(this.socket_url);
+
+          this.socket.on('devfest', (req) => this._handleRequest(req));
+          this.socket.on('close', (req) => this._showEverything());
           
           this.planning = false;
           this.displayCurrentTalk = false;
@@ -100,7 +106,7 @@
           this.currentTalks = this.talks.filter(talk => {
             return talk.title.indexOf(talkTitle) != -1
           });
-          console.log(this.currentTalks)
+
           this._hideEverything();
           this.displayCurrentTalk = true;
         }
@@ -159,7 +165,7 @@
         }
 
         _displayBio (name) {
-          this.speaker = this.speakers.filter(speaker => speaker.name === name)[0];
+          this.speaker = this.speakers.filter(speaker => speaker.name.indexOf(name) != -1)[0];
         }
 
         _fetchDevfestTalks () {
