+++ speechback/config.example.js
@@ -1,0 +1,7 @@
+// API.AI
+module.exports.apiaitoken = 'YOUR_API_AI_TOKEN';
+
+// MICROPHONE
+module.exports.microphone = {
+  threshold: 0
+};
