+++ speechback/package.json
@@ -16,7 +16,6 @@
   },
   "dependencies": {
     "apiai-promise": "^1.0.0",
-    "ip": "^1.1.5",
     "node-record-lpcm16": "^0.3.0",
     "snowboy": "^1.2.0",
     "socket.io": "^2.0.1"
