+++ speechback/resources/common.res
