--- tmpback/config.example.js
@@ -1,11 +1,0 @@
-// API.AI
-module.exports.apiaitoken = 'YOUR_API_AI_TOKEN';
-
-// GITHUB
-module.exports.github_client_id = 'YOUR_GITHUB_CLIENT_ID'
-module.exports.github_client_secret = 'YOUR_GITHUB_CLIENT_SECRET';
-
-// MICROPHONE
-module.exports.microphone = {
-  threshold: 0
-};
